refactor(table): [SIDE-399] Swap Section and Contents
