--- src/lib/components/table/components/TableContent/TableContent.module.scss
@@ -1,15 +1,0 @@
-@use '../../../../scss/public/colors' as *;
-@use '../../../../scss/public/grid' as *;
-
-.table {
-  border-collapse: collapse;
-  table-layout: fixed;
-
-  @include p-size-tablet {
-    min-width: 845px;
-  }
-}
-
-.tr {
-  min-height: 72px;
-}
