+++ src/lib/components/table/components/TableContents/TableContents.module.scss
@@ -1,0 +1,10 @@
+.cardWrapper {
+  height: 88px;
+}
+
+.card {
+  position: absolute;
+  left: 0;
+  right: 0;
+  z-index: 3;
+}
